Crate shlex

Same idea as (but implementation not directly based on) the Python shlex module. However, this implementation does not support any of the Python module's customization because it makes parsing slower and is fairly useless. You only get the default settings of shlex.split, which mimic the POSIX shell: http://pubs.opengroup.org/onlinepubs/9699919799/utilities/V3_chap02.html

This implementation also deviates from the Python version in not treating \r specially, which I believe is more compliant.

The algorithms in this crate are oblivious to UTF-8 high bytes, so they iterate over the bytes directly as a micro-optimization.

Structs

Shlex
An iterator that takes an input string and splits it into the words using the same syntax as the POSIX shell.

Functions

quote
Given a single word, return a string suitable to encode it as a shell argument.

split
Convenience function that consumes the whole string at once. Returns None if the input was erroneous.